Added workPage.tsx to route folder
I am currently working through how to make the workPage show up. I added it as an element under "children" in main.tsx, but there seems to be a type error? (refer to line 19 in main.tsx). The error only shows here and not on workPage.tsx.

Added:

src > route > workPage.tsx